practice-1-legacy-app/broken-weather-app/requirements.txt
Flask==2.0.1
psycopg2-binary==2.8.6
redis==3.5.3
requests==2.25.1
flask-cors==3.0.10
configparser==5.0.2
numpy==1.19.5
Pillow==8.2.0
pytz==2021.1
Werkzeug==2.0.1
gunicorn